Move snapshot utils to own file
